+++ packages/cheqd-blockchain-modules/tests/offchain-signatures-module.test.js
@@ -1,17 +1,17 @@
-import { CheqdAPI } from "@docknetwork/cheqd-blockchain-api";
+import { CheqdAPI, CheqdNetwork } from "@docknetwork/cheqd-blockchain-api";
 import {
   CheqdTestnetDid,
   CheqdOffchainSignatureParamsRef,
 } from "@docknetwork/credential-sdk/types";
 import { MultiApiOffchainSignaturesModule } from "@docknetwork/credential-sdk/modules";
 import generateOffchainSignaturesModuleTests from "@docknetwork/credential-sdk/modules/tests/offchain-signatures-module";
-import { faucet, url } from "./constants";
+import { MultiApiDIDModule, MultiApiCoreModules } from "@docknetwork/credential-sdk/modules";
+import { faucet, url, network } from "./constants";
 import {
   CheqdOffchainSignaturesModule,
   CheqdDIDModule,
   CheqdCoreModules
 } from "../src";
-import { MultiApiDIDModule, MultiApiCoreModules } from "@docknetwork/credential-sdk/modules";
 import CheqdBBSModule from "../src/offchain-signatures/bbs";
 import CheqdBBSPlusModule from "../src/offchain-signatures/bbs-plus";
 import CheqdPSModule from "../src/offchain-signatures/ps";
@@ -22,8 +22,8 @@ describe("OffchainSignaturesModule", () => {
   beforeAll(async () => {
     await cheqd.init({
       url,
-      mnemonic: faucet.mnemonic,
-      network: "testnet",
+      wallet: await faucet.wallet(),
+      network,
     });
   });