+++ packages/credential-sdk/src/utils/interfaces.js
@@ -1,11 +1,17 @@
 /**
- * Attempts to call `value.toJSON()`, returns `value` if method doesn't exist.
+ * Attempts to call `value.toJSON()`, returns `JSON.parse(JSON.stringify(value))` if method doesn't exist.
  * @template T
  * @param {T} value
- * @returns {T | Object}
+ * @returns {object}
  */
-export const maybeToJSON = (value) => (value && typeof value.toJSON === 'function' ? value.toJSON() : value);
+export const maybeToJSON = (value) => (value && typeof value.toJSON === 'function' ? value.toJSON() : JSON.parse(JSON.stringify(value)));
 
+/**
+ * Attempts to call `value.toJSON()` and stringify the result, returns `String(value)` in case of failure.
+ * @template T
+ * @param {T} value
+ * @returns {string}
+ */
 export const maybeToJSONString = (value) => {
   const json = maybeToJSON(value);
 
@@ -29,15 +35,15 @@ export const maybeEq = (value, other) => (value && typeof value.eq === 'function
  * Attempts to call `value.toHuman()` or `value.toJSON`, returns `value` if methods don't exist.
  * @template T
  * @param {T} value
- * @returns {T | Object}
+ * @returns {object}
  */
 export const maybeToHuman = (obj) => (obj && typeof obj.toHuman === 'function' ? obj.toHuman() : maybeToJSON(obj));
 
 /**
- * Attempts to call `value.toNumber()`, returns `value` if method doesn't exist.
+ * Attempts to call `value.toNumber()`, returns `+value` if method doesn't exist.
  * @template T
  * @param {T} value
- * @returns {T | number}
+ * @returns {number}
  */
 export const maybeToNumber = (value) => (value && typeof value.toNumber === 'function' ? value.toNumber() : +value);
 
@@ -52,7 +58,6 @@ export const NotAConstructor = Symbol.for(
  * Attempts to intantiate new object of the supplied class using provided arguments.
  * @param Class
  * @param args
- * @returns {T | number}
  */
 export const maybeNew = (Class, args) => (!Class[NotAConstructor] ? new Class(...args) : Class.apply(Class, args));
 
@@ -60,7 +65,6 @@ export const maybeNew = (Class, args) => (!Class[NotAConstructor] ? new Class(..
  * Attempts to create new instance of the supplied class using `Class.from(obj)`, instantiates class if `from` doesn't exist.
  * @param Class
  * @param args
- * @returns {T | number}
  */
 export const maybeFrom = (klass, obj) => (typeof klass.from === 'function' ? klass.from(obj) : maybeNew(klass, [obj]));